refactor(streaming): split Pi completion profile
